+++ musictodmx/fire.h
@@ -9,11 +9,11 @@
 #define FIRE_BUTTON_PIN 14
 #define FIRE_DURATION   34 //flame's duration [0-255]
 
-extern uint8_t fire_channel;
-extern uint8_t fire_channel_duration;
+extern uint16_t fire_channel;
+extern uint16_t fire_channel_duration;
 extern TaskHandle_t taskFireHandle;
 
-extern void init_fire(uint8_t channel);
+extern void init_fire(uint16_t channel);
 extern void fireHandler(void);
 extern void fireSelector();
